Initial protobufs for events
Rather than the ad-hoc string serialization used for the prototype,
we'll begin to evolve a Protobuf-based encoding for our events.

The current thinking is that all events will have a "supervisor
metadata" field. Any events that pertain to services will have an
additional "service metadata" field. Any additional information will
be added as fields on the specific concrete event.

For creating events, we'll use a functional interface at the sites
where events are created. These functions then initialize the
protobuf-derived Rust types according to their arguments. When we go
to publish the events, we'll inject the (global) Supervisor metadata,
serialize the whole thing to bytes, and send them to NATS.

Signed-off-by: Christopher Maier <cmaier@chef.io>