Add the IP address of the Supervisor to the event metadata
Currently, this is the address on which the supervisor listens for
gossip. We may want to change this in the future, but it's not 100%
clear how we might best do this, since the Supervisor doesn't
technically need to be listening for gossip, control commands, or HTTP
gateway requests on any externally-accessible address.

Signed-off-by: Christopher Maier <cmaier@chef.io>