Consolidate error handling for event subsystem
Instead of adding a bunch of error cases to the top-level `Error`
enum, we'll create one specifically for event-related errors. That
way, all the implementation details stay within the event subsystem,
and we only have to translate one type into the top-level `Error`,
rather than do it on a case-by-case basis.

This is a little experimental at this point, but I think it could lead
to better error handling overall. For instance, the event-related
errors are more descriptive of what's actually going on in the domain,
rather than serving as simple wrappers around dependency errors. Those
errors are still present, but exposed through the `source` method.

Signed-off-by: Christopher Maier <cmaier@chef.io>