Run health checks asynchronously
Previously, health check hooks were run synchronously by the
Supervisor, with logic for managing the scheduling of the next
invocation at some time in the future. Each time through the service's
"tick", we would have to determine whether or not to run a health
check based on this. Not only does this add complexity, but it also
ends up blocking the Supervisor if a health check hook takes any
appreciable amount of time.

Here, we reformulate health checking to happen in an endlessly
repeating future that runs asynchronously. Rather than tracking when
the next time to run a hook is, this is now baked into the logic of
the future itself, giving us less to explicitly manage.

Now, when we start a service, we spawn the health checking future. It
executes the health check, then waits for the appropriate amount of
time; these two steps constitute a single iteration of the overall
health checking process. This iteration is then repeated endlessly.

The future is constructed such that we retain a "handle" to it, which
can be used to terminate the future when, e.g., the service shuts
down. This handle is the only explicit "state" that the Service itself
needs to keep track of, and even then, all it really has to do is just
maintain ownership of it.

In order to make this approach work with how Services are currently
constructed, we needed to wrap the `HealthCheckHook` itself in an
`Arc`, as well as the service's process supervisor and health check
result fields. As we continue to move toward a more asynchronous
internal architecture, we hope to be able to eliminate this shared
state.

Signed-off-by: Christopher Maier <cmaier@chef.io>